Extract jsonrpc core types into trin-types workspace
