+++ ethportal-api/Cargo.toml
@@ -11,21 +11,31 @@ categories = ["cryptography::cryptocurrencies"]
 authors = ["https://github.com/ethereum/trin/graphs/contributors"]
 
 [dependencies]
+anyhow = "1.0.68"
 bytes = "1.3.0"
+discv5 = { version = "0.2.1", features = ["serde"]}
 enr = {version= "0.7.0", features=["serde", "ed25519"]}
 ethereum-types = "0.12.1"
 eth2_ssz = "0.4.0"
 eth2_ssz_derive = "0.3.0"
 eth2_ssz_types = "0.2.1"
 jsonrpsee = {version="0.16.2", features = ["async-client", "client", "macros", "server"]}
-reth-primitives = { version = "0.1.0", git = "https://github.com/paradigmxyz/reth.git" }
-reth-rlp = { version = "0.1.2", git = "https://github.com/paradigmxyz/reth.git" }
+reth-primitives = {git = "https://github.com/paradigmxyz/reth.git", rev = "161795945b21969faa06cd1234a0b5e3bf62f530"}
+reth-rlp = {git = "https://github.com/paradigmxyz/reth.git", rev = "161795945b21969faa06cd1234a0b5e3bf62f530"}
 serde = { version = "1.0.150", features = ["derive"] }
 serde_json = "1.0.89"
 serde-hex = "0.1.0"
 sha2 = "0.10.1"
 thiserror = "1.0.38"
+trin-types = { path = "../trin-types" }
 trin-utils = { path = "../trin-utils" }
+tokio = { version = "1.14.0", features = ["full"] }
+ureq = { version = "2.5.0", features = ["json"] }
+validator = { version = "0.13.0", features = ["derive"] }
 
 [dev-dependencies]
+env_logger = "0.9.0"
 rstest = "0.16.0"
+test-log = { version = "0.2.11", features = ["trace"] }
+tracing = "0.1.36"
+tracing-subscriber = "0.3.15"
