+++ ethportal-api/src/types/mod.rs
@@ -1,4 +1,6 @@
 pub mod content_item;
 pub mod content_key;
 pub mod discv5;
+pub mod params;
 pub mod portal;
+pub mod request;
