+++ ethportal-api/src/types/params.rs
@@ -1,14 +1,8 @@
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
-use tokio::sync::mpsc;
-use validator::{Validate, ValidationError};
+use validator::ValidationError;
 
-use crate::{
-    jsonrpc::endpoints::{HistoryEndpoint, StateEndpoint, TrinEndpoint},
-    portalnet::types::messages::SszEnr,
-};
-
-type Responder<T, E> = mpsc::UnboundedSender<Result<T, E>>;
+use trin_types::enr::SszEnr;
 
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
@@ -30,49 +24,6 @@ impl From<Params> for Value {
             Params::None => Value::Null,
         }
     }
-}
-
-#[derive(Debug, Deserialize, Serialize, Validate, Clone)]
-pub struct JsonRequest {
-    #[validate(custom = "validate_jsonrpc_version")]
-    pub jsonrpc: String,
-    #[serde(default = "default_params")]
-    pub params: Params,
-    pub method: String,
-    pub id: u32,
-}
-
-// Global portal network JSON-RPC request
-#[derive(Debug, Clone)]
-pub struct PortalJsonRpcRequest {
-    pub endpoint: TrinEndpoint,
-    pub resp: Responder<Value, anyhow::Error>,
-    pub params: Params,
-}
-
-/// History network JSON-RPC request
-#[derive(Debug, Clone)]
-pub struct HistoryJsonRpcRequest {
-    pub endpoint: HistoryEndpoint,
-    pub resp: Responder<Value, String>,
-}
-
-/// State network JSON-RPC request
-#[derive(Debug)]
-pub struct StateJsonRpcRequest {
-    pub endpoint: StateEndpoint,
-    pub resp: Responder<Value, String>,
-}
-
-fn default_params() -> Params {
-    Params::None
-}
-
-fn validate_jsonrpc_version(jsonrpc: &str) -> Result<(), ValidationError> {
-    if jsonrpc != "2.0" {
-        return Err(ValidationError::new("Unsupported jsonrpc version"));
-    }
-    Ok(())
 }
 
 #[derive(Debug)]
@@ -107,30 +58,6 @@ impl TryFrom<&Value> for NodesParams {
 mod test {
     use super::*;
     use rstest::rstest;
-    use validator::ValidationErrors;
-
-    #[test_log::test]
-    fn test_json_validator_accepts_valid_json() {
-        let request = JsonRequest {
-            jsonrpc: "2.0".to_string(),
-            id: 1,
-            params: Params::None,
-            method: "eth_blockNumber".to_string(),
-        };
-        assert_eq!(request.validate(), Ok(()));
-    }
-
-    #[test_log::test]
-    fn test_json_validator_with_invalid_jsonrpc_field() {
-        let request = JsonRequest {
-            jsonrpc: "1.0".to_string(),
-            id: 1,
-            params: Params::None,
-            method: "eth_blockNumber".to_string(),
-        };
-        let errors = request.validate();
-        assert!(ValidationErrors::has_error(&errors, "jsonrpc"));
-    }
 
     fn expected_map() -> Map<String, Value> {
         let mut expected_map = serde_json::Map::new();
