+++ ethportal-peertest/Cargo.toml
@@ -32,6 +32,7 @@ trin = { path = ".." }
 trin-core = { path = "../trin-core" }
 trin-history = { path = "../trin-history" }
 trin-state = { path = "../trin-state" }
+trin-types = { path = "../trin-types" }
 trin-utils = { path = "../trin-utils" }
 ureq = { version = "2.5.0", features = ["json"] }
