+++ newsfragments/588.fixed.md
@@ -1,0 +1,1 @@
+Extracted trin-core/jsonrpc/types into trin-types workspace
