+++ trin-core/Cargo.toml
@@ -14,7 +14,6 @@ authors = ["https://github.com/ethereum/trin/graphs/contributors"]
 anyhow = "1.0.68"
 async-recursion = "1.0.0"
 async-trait = "0.1.53"
-base64 = "0.13.0"
 bytes = "1.3.0"
 clap = "2.33.3"
 delay_map = "0.1.1"
@@ -55,10 +54,11 @@ tokio = { version = "1.14.0", features = ["full"] }
 tracing = "0.1.36"
 tree_hash = "0.4.0"
 tree_hash_derive = "0.4.0"
+trin-types = { path="../trin-types" }
 trin-utils = { path="../trin-utils" }
 ureq = { version = "2.5.0", features = ["json"] }
-validator = { version = "0.13.0", features = ["derive"] }
 url = "2.3.1"
+validator = { version = "0.13.0", features = ["derive"] }
 
 [target.'cfg(windows)'.dependencies]
 ipconfig = "0.2.2"
