--- trin-core/src/jsonrpc/mod.rs
@@ -1,4 +1,0 @@
-pub mod endpoints;
-pub mod service;
-pub mod types;
-pub mod utils;
