+++ trin-core/src/lib.rs
@@ -2,7 +2,6 @@
 #![warn(clippy::cargo)]
 
 pub mod cli;
-pub mod jsonrpc;
 pub mod locks;
 pub mod portalnet;
 pub mod socket;
