+++ trin-core/src/portalnet/mod.rs
@@ -1,7 +1,5 @@
 #![allow(clippy::all)]
 
-use discv5::enr::CombinedKey;
-
 pub mod discovery;
 pub mod events;
 pub mod find;
@@ -10,5 +8,3 @@ pub mod overlay;
 mod overlay_service;
 pub mod storage;
 pub mod types;
-
-pub type Enr = discv5::enr::Enr<CombinedKey>;
