+++ trin-history/Cargo.toml
@@ -23,6 +23,7 @@ tokio = { version = "1.14.0", features = ["full"] }
 tracing = "0.1.36"
 tree_hash = "0.4.0"
 trin-core = { path = "../trin-core" }
+trin-types = { path = "../trin-types" }
 trin-utils = { path = "../trin-utils" }
 
 [dev-dependencies]
