+++ trin-history/src/network.rs
@@ -10,13 +10,13 @@ use trin_core::{
         storage::{PortalStorage, PortalStorageConfig},
         types::{
             content_key::HistoryContentKey,
-            distance::XorMetric,
             messages::{PortalnetConfig, ProtocolId},
         },
     },
     types::validation::HeaderOracle,
     utp::stream::UtpListenerRequest,
 };
+use trin_types::distance::XorMetric;
 
 use crate::validation::ChainHistoryValidator;
