+++ trin-state/src/lib.rs
@@ -11,8 +11,8 @@ use tokio::{
 use tracing::info;
 
 use crate::{events::StateEvents, jsonrpc::StateRequestHandler};
+use ethportal_api::types::request::StateJsonRpcRequest;
 use trin_core::{
-    jsonrpc::types::StateJsonRpcRequest,
     portalnet::{
         discovery::Discovery, storage::PortalStorageConfig, types::messages::PortalnetConfig,
     },
