+++ trin-types/Cargo.toml
@@ -1,31 +1,30 @@
 [package]
-name = "trin-state"
+name = "trin-types"
 version = "0.1.0"
 edition = "2021"
-repository = "https://github.com/ethereum/trin/tree/master/trin-state"
+repository = "https://github.com/ethereum/trin/tree/main/trin-types"
 license = "GPL-3.0"
 readme = "README.md"
 keywords = ["ethereum", "portal-network"]
 categories = ["cryptography::cryptocurrencies"]
-description = "State network subprotocol for Trin."
+description = "Types library for Trin."
 authors = ["https://github.com/ethereum/trin/graphs/contributors"]
 
 [dependencies]
-anyhow = "1.0.68"
-async-trait = "0.1.53"
+base64 = "0.13.0"
 discv5 = { version = "0.2.1", features = ["serde"]}
 ethereum-types = "0.12.1"
-ethportal-api = { path = "../ethportal-api" }
-eth_trie = "0.1.0"
-num = "0.4.0"
-parking_lot = "0.11.2"
-tracing = "0.1.36"
-tokio = {version = "1.14.0", features = ["full"]}
-trin-core = { path = "../trin-core" }
-trin-types = { path = "../trin-types" }
-rocksdb = "0.18.0"
+eth2_ssz = "0.4.0"
+eth2_ssz_derive = "0.3.0"
+eth2_ssz_types = "0.2.1"
+rlp = "0.5.0"
+serde_json = "1.0.89"
+validator = { version = "0.13.0", features = ["derive"] }
 
 [dev-dependencies]
 env_logger = "0.9.0"
+quickcheck = "1.0.3"
+quickcheck_macros = "1.0.0"
 test-log = { version = "0.2.11", features = ["trace"] }
+tracing = "0.1.36"
 tracing-subscriber = "0.3.15"
