+++ trin-types/README.md
@@ -1,0 +1,2 @@
+# Trin-Types
+Library for shared types across trin workspace members. Types only used by a single workspace member belong in a root-level `types` module.
