+++ utp-testing/Cargo.toml
@@ -19,6 +19,7 @@ structopt = "0.3.26"
 tracing = "0.1.36"
 tracing-subscriber = "0.3.15"
 trin-core = { path = "../trin-core" }
+trin-types = { path = "../trin-types" }
 trin-utils = { path = "../trin-utils" }
 tokio = {version = "1.14.0", features = ["full"]}
